Rename operator to operator.controller, replace KubeCluster shim with deprecation warning, move new KubeCluster to operator
