+++ dask_kubernetes/operator/__init__.py
@@ -1,1 +1,7 @@
-from .operator import *
+from .kubecluster import (
+    KubeCluster,
+    make_cluster_spec,
+    make_scheduler_spec,
+    make_worker_spec,
+    discover,
+)
