+++ dask_kubernetes/operator/controller/tests/resources/simpleworkergroup.yaml
