+++ dask_kubernetes/operator/kubecluster/tests/conftest.py
@@ -2,7 +2,7 @@ import pytest
 
 import dask.config
 
-from dask_kubernetes.operator.kubecluster import KubeCluster
+from dask_kubernetes.operator import KubeCluster
 
 
 @pytest.fixture
