+++ setup.py
@@ -22,8 +22,8 @@ setup(
     entry_points="""
         [dask_cluster_discovery]
         helmcluster=dask_kubernetes.helm:discover
-        kubecluster=dask_kubernetes.operator.kubecluster:discover
+        kubecluster=dask_kubernetes.operator:discover
         [dask_operator_plugin]
-        noop=dask_kubernetes.operator.plugins.noop
+        noop=dask_kubernetes.operator.controller.plugins.noop
       """,
 )
